requirements.txt
Flask==2.3.3
Flask-WTF==1.1.1
Pillow==10.0.0
matplotlib==3.7.2
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0